Fine tune YouthProfile RENEWING membership status logic
While a youth profile is being renewed and it's still waiting
for an approval from a guardian, the RENEWING status is returned
for a profile which is still active while the renewal is ongoing.
PENDING will be returned for an expired youth profile, which is
waiting for the approval from a guardian.

Different status now are:
- ACTIVE = membership is active
- PENDING = membership is waiting for approval and is either inactive or expired
- RENEWING = membership is active, but renewal is waiting for approval
- EXPIRED = membership has expired

Refs YM-463, YM-469