Profile is not renewable if current expiration is greater than what a renewal would yield
This makes sure a profile is not shown to be renewable e.g. if
YOUTH_MEMBERSHIP_SEASON_END_DATE is changed to be earlier than what it currently is, since
a renewal in this case would mean the expiration date goes backwards.

Also it will allow us to have users with a RENEWING status in staging, when
a renewal will not yet yield a full season renewal.

Refs YM-463, YM-469